configure dns server on azure bridge for ubuntu versions > 18. Moved hns related apis from platform package to hsnclient package
