+++ common/utils.go
@@ -4,8 +4,11 @@
 package common
 
 import (
+	"bufio"
 	"encoding/binary"
 	"encoding/xml"
+	"fmt"
+	"io"
 	"net"
 	"os"
 
@@ -122,4 +125,33 @@ func StartProcess(path string, args []string) error {
 	}
 
 	return err
+}
+
+// ReadFileByLines reads file line by line and return array of lines.
+func ReadFileByLines(filename string) ([]string, error) {
+	var (
+		lineStrArr []string
+	)
+
+	f, err := os.Open(filename)
+	if err != nil {
+		return nil, fmt.Errorf("Error opening %s file error %v", filename, err)
+	}
+
+	defer f.Close()
+
+	r := bufio.NewReader(f)
+
+	for {
+		lineStr, err := r.ReadString('\n')
+		if err != nil {
+			if err != io.EOF {
+				return nil, fmt.Errorf("Error reading %s file error %v", filename, err)
+			}
+			break
+		}
+		lineStrArr = append(lineStrArr, lineStr)
+	}
+
+	return lineStrArr, nil
 }
