+++ platform/os_windows.go
@@ -5,15 +5,12 @@ package platform
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"os/exec"
 	"strings"
 	"time"
 
-	"github.com/Azure/azure-container-networking/cns"
 	"github.com/Azure/azure-container-networking/log"
-	"github.com/Microsoft/hcsshim"
 )
 
 const (
@@ -199,131 +196,6 @@ func SetSdnRemoteArpMacAddress() error {
 	return nil
 }
 
-// CreateHnsNetwork creates the HNS network with the provided configuration
-func CreateHnsNetwork(nwConfig cns.CreateHnsNetworkRequest) error {
-	log.Printf("[Azure CNS] CreateHnsNetwork")
-	// Initialize HNS network.
-	hnsNetwork := &hcsshim.HNSNetwork{
-		Name:                 nwConfig.NetworkName,
-		Type:                 nwConfig.NetworkType,
-		NetworkAdapterName:   nwConfig.NetworkAdapterName,
-		SourceMac:            nwConfig.SourceMac,
-		DNSSuffix:            nwConfig.DNSSuffix,
-		DNSServerList:        nwConfig.DNSServerList,
-		DNSServerCompartment: nwConfig.DNSServerCompartment,
-		ManagementIP:         nwConfig.ManagementIP,
-		AutomaticDNS:         nwConfig.AutomaticDNS,
-	}
-
-	for _, policy := range nwConfig.Policies {
-		hnsNetwork.Policies = append(hnsNetwork.Policies, policy)
-	}
-
-	for _, subnet := range nwConfig.Subnets {
-		hnsSubnet := hcsshim.Subnet{
-			AddressPrefix:  subnet.AddressPrefix,
-			GatewayAddress: subnet.GatewayAddress,
-		}
-
-		hnsNetwork.Subnets = append(hnsNetwork.Subnets, hnsSubnet)
-	}
-
-	for _, macPool := range nwConfig.MacPools {
-		hnsMacPool := hcsshim.MacPool{
-			StartMacAddress: macPool.StartMacAddress,
-			EndMacAddress:   macPool.EndMacAddress,
-		}
-		hnsNetwork.MacPools = append(hnsNetwork.MacPools, hnsMacPool)
-	}
-
-	return createHnsNetwork(hnsNetwork)
-}
-
-// DeleteHnsNetwork deletes the HNS network with the provided name
-func DeleteHnsNetwork(networkName string) error {
-	log.Printf("[Azure CNS] DeleteHnsNetwork")
-
-	return deleteHnsNetwork(networkName)
-}
-
-// CreateDefaultExtNetwork creates default HNS network named ext (if it doesn't exist already)
-// to create external switch on windows platform.
-// This allows orchestrators to start CNS which pre-provisions the network so that the
-// VM network blip / disconnect is avoided when calling cni add for the very first time.
-func CreateDefaultExtNetwork(networkType string) error {
-	networkType = strings.ToLower(strings.TrimSpace(networkType))
-	if len(networkType) == 0 {
-		return nil
-	}
-
-	if networkType != hnsL2Bridge && networkType != hnsL2Tunnel {
-		return fmt.Errorf("Invalid hns network type %s", networkType)
-	}
-
-	log.Printf("[Azure CNS] CreateDefaultExtNetwork")
-	extHnsNetwork, _ := hcsshim.GetHNSNetworkByName(ExtHnsNetworkName)
-
-	if extHnsNetwork != nil {
-		log.Printf("[Azure CNS] Found existing DefaultExtNetwork with type: %s", extHnsNetwork.Type)
-		if !strings.EqualFold(networkType, extHnsNetwork.Type) {
-			return fmt.Errorf("Network type mismatch with existing network: %s", extHnsNetwork.Type)
-		}
-
-		return nil
-	}
-
-	// create new hns network
-	log.Printf("[Azure CNS] Creating DefaultExtNetwork with type %s", networkType)
-
-	hnsNetwork := &hcsshim.HNSNetwork{
-		Name: ExtHnsNetworkName,
-		Type: networkType,
-	}
-
-	hnsSubnet := hcsshim.Subnet{
-		AddressPrefix:  ExtHnsNetworkAddressPrefix,
-		GatewayAddress: ExtHnsNetworkGwAddress,
-	}
-
-	hnsNetwork.Subnets = append(hnsNetwork.Subnets, hnsSubnet)
-
-	return createHnsNetwork(hnsNetwork)
-}
-
-// DeleteDefaultExtNetwork deletes the default HNS network
-func DeleteDefaultExtNetwork() error {
-	log.Printf("[Azure CNS] DeleteDefaultExtNetwork")
-
-	return deleteHnsNetwork(ExtHnsNetworkName)
-}
-
-// createHnsNetwork calls the hcshim to create the hns network
-func createHnsNetwork(hnsNetwork *hcsshim.HNSNetwork) error {
-	// Marshal the request.
-	buffer, err := json.Marshal(hnsNetwork)
-	if err != nil {
-		return err
-	}
-	hnsRequest := string(buffer)
-
-	// Create the HNS network.
-	log.Printf("[Azure CNS] HNSNetworkRequest POST request:%+v", hnsRequest)
-	hnsResponse, err := hcsshim.HNSNetworkRequest("POST", "", hnsRequest)
-	log.Printf("[Azure CNS] HNSNetworkRequest POST response:%+v err:%v.", hnsResponse, err)
-
-	return err
-}
-
-// deleteHnsNetwork calls HNS to delete the network with the provided name
-func deleteHnsNetwork(networkName string) error {
-	hnsNetwork, err := hcsshim.GetHNSNetworkByName(networkName)
-	if err == nil {
-		// Delete the HNS network.
-		var hnsResponse *hcsshim.HNSNetwork
-		log.Printf("[Azure CNS] HNSNetworkRequest DELETE id:%v", hnsNetwork.Id)
-		hnsResponse, err = hcsshim.HNSNetworkRequest("DELETE", hnsNetwork.Id, "")
-		log.Printf("[Azure CNS] HNSNetworkRequest DELETE response:%+v err:%v.", hnsResponse, err)
-	}
-
-	return err
+func GetOSDetails() (map[string]string, error) {
+	return nil, nil
 }
